+++ gframe/deck_manager.cpp
@@ -172,7 +172,7 @@ DeckError DeckManager::CheckDeck(Deck& deck, int lfhash, DuelAllowedCards allowe
 	if(TypeCount(deck.main, forbiddentypes) > 0 || TypeCount(deck.extra, forbiddentypes) > 0 || TypeCount(deck.side, forbiddentypes) > 0)
 		return ret.type = DeckError::FORBTYPE, ret;
 	bool speed = mainGame->extra_rules & DECK_LIMIT_20;
-	int minmain = 40, maxmain = 60, maxextra = 15, maxside = 15;
+	size_t minmain = 40, maxmain = 60, maxextra = 15, maxside = 15;
 	if(doubled){
 		if(speed){
 			maxextra = 10;
